api: introduce a testing util for dom
