+++ tests/browser.rs
@@ -1,5 +1,6 @@
 #![allow(unused_variables)]
 use hirola::prelude::*;
+use hirola_core::dom_test_utils::next_tick;
 use wasm_bindgen::{prelude::Closure, JsCast};
 use wasm_bindgen_test::*;
 use web_sys::{Element, Node};
@@ -15,15 +16,6 @@ fn body() -> Node {
 fn inner_html(element: &Node) -> String {
     let element = element.dyn_ref::<Element>().unwrap();
     element.inner_html()
-}
-
-#[allow(dead_code)]
-fn next_tick<F: Fn() + 'static>(f: F) {
-    let a = Closure::<dyn Fn()>::new(move || f());
-    web_sys::window()
-        .unwrap()
-        .set_timeout_with_callback(a.as_ref().unchecked_ref())
-        .unwrap();
 }
 
 #[wasm_bindgen_test]
